Feat: rozwiązanie przez ziarno świadome kontekstu i wyciągające zeń ziarno do późniejszej inicjacji. Hak. Uwaga - mamy nieskończoną pętlę wzajemnych wywołań między ziarnami
Detal: ignoruję pliki .attach_pid wynikłe z drobnego niedociągnięcia / błędu w IDE.

There are two ways of constructing a software design.  One way is to make it so simple that there are obviously no deficiencies. And the other way is to make it so complicated that there are no obvious deficiencies.
    -- C.A.R. Hoare